Extract NodeInfo creation
